refac(chatWindow): convert ChatWindow component to TypeScript and improve typing
